+++ src/Ability.ts
@@ -30,7 +30,6 @@ export enum Ability {
     Shock = "Shock", // "the first time you strike with a crit, also discard one of opponent's unresolved norms (or crit if no norms)"
 
     // fight stuff relevant to only 2021 or I have not checked for 2024
-    Stun2021 = "Stun2021", // effect different between shoot and fight
     StormShield2021 = "StormShield2021", // each parry cancels two successes
     Hammerhand2021 = "Hammerhand2021", // first strike does one extra dmg
     Dueller = "Dueller2021", // each crit parry cancels extra normal success